requirements.txt
joblib
mlflow
mlxtend
numpy
pandas
PyYAML
scikit_learn
streamlit
